scripting: Clingon trick to pass unknown options to the script without "--"
a caveat: it does not catch legitimate unknown options given before
the script name,
and it catches known options given after it, though we would like not
to.

$ ./script.lisp -v

-v is caught for CIEL.

$ ./script.lisp -- -v

-v is passed to script.

$ ciel -x -s script

-x is not caught as unknown for CIEL but passed along to script.

$ ciel -v -x -s script

the -v at first position allows to see that -x is passed along.

$ ./script.lisp -x 4242

for other cases, we don't need a "--" and that's pretty cool.

Damn, was it worth spending a day on it?