terminal repl: re-add use of external terminal. Not yet useful for Slime.
The "!" prefix works in Slime, it comes from the Clesh shell
passthrough.
On our repl we catch earlier if the command is a visual command, and
run it approprietaly.

We should contribute this to Clesh to have an equivalent behaviour in
Slime (except we would open a terminal window, or in Emacs' vterm).

for #71